Split out deletePost
